Supported Transports
The current version of Protocol.md states this:

> ReactiveSocket as specified here only allows for TCP, WebSocket, and Aeron as transport protocols.

This is at https://github.com/ReactiveSocket/reactivesocket/blob/master/Protocol.md#transport-protocol

Then in the next section, a couple paragraph away it mentions HTTP/2 as needing frame length.

Link: https://github.com/ReactiveSocket/reactivesocket/blob/master/Protocol.md#framing-protocol-usage

It is confusing to say the protocol only supports TCP, WebSocket, and Aeron, and then mention "HTTP/2" and "Other" in the framing section. 

In this change I propose a possible wording to resolve the confusion. But I'm not quite sure what was meant by this sentence, when the section itself is also permitting other transports. 

So what is trying to be said, and what can we do to make this clearer?