Other Transport Protocols and Framing
Why does this table show all other transports as "No"?

That suggests that a user of the protocol can never take advantage of this optimization unless it is WebSocket or Aeron. Why do we need to restrict it? Especially when in the previous section we gave the requirements for a transport protocol.