64-bit Trace Id
Previously the extension specification assumed that all trace ids would be
128-bit values.  In the discussion of #256 it was highlighted that not only
was this variable between 64-bit and 128-bit, but the primary behavior is
64-bit values.  This change updates the specification to add a new T flag that
indicates whether the value is a 64-bit value (unset) or a 128-bit value
(set).  This removes the need to pad 64-bit values and ensures the efficiency
in the data structure.

Signed-off-by: Ben Hale <bhale@pivotal.io>