Provides addition to QoS and Prioritisation (#280)
* Provides addition to QoS and Prioritisation

This PR provides an update to the spec, which mentions a Recommendation for all implementations. Basically, it adds a note on prioritized delivering of all the *Stream ID 0* frames in case it is possible.

For example, in RSocket-Java, we have a Queue, which basically enqueues all the frames on the Requester level first, and the underlying duplex connection drains them and delivers later on. It means that the KeepAlive frame as an important one could be stacked in the `Queue` and delayed in its delivery to the connection.

Undelivered on time KeepAlive frame could cause unwanted connection close event, which in that case is a false-positive action.

The same relates to Lease, Metadata_Push frames which brings emerging information about peers connected via the network